New approach to get active interfaces on Linux
Since the output of "ifconfig" has changed, a new approach is necessary. The new method first searches for all lines that start with a character between a and z, thus omitting the blank lines between interfaces. An example of what would be returned is "enp2s0: flags=4163<UP,BROADCAST,RUNNING,MULTICAST>  mtu 1500". Next, this line is split by the space character and uses the first column for further processing, so in our example that would be "enp2s0:". Lastly, the ":" at the end is being removed.